requirements.txt
pandas==3.0.1
pyarrow==23.0.1
google-cloud-bigquery==3.34.0
pytz==2025.2
openpyxl==3.1.5
apache-airflow==2.10.4
apache-airflow-providers-google==10.22.0
apache-airflow-providers-docker==3.14.0
python-dotenv==1.0.1
